Fix for jekyll/jekyll@0f74db4

+++ lib/jekyll-minifier.rb
@@ -77,6 +77,10 @@ module Jekyll
 
     def write(dest)
       dest_path = destination(dest)
+
+      return false if File.exist?(dest_path) and !modified?
+      self.class.mtimes[path] = mtime
+
       if exclude?(dest, dest_path)
         copy_file(path, dest_path)
       else
